refactoring redux data to integrate with react components
