+++ expensify/redux.js
@@ -1,115 +1,9 @@
-import { createStore, combineReducers } from 'redux';
-import uuid from 'uuid';
 
-const defaultExpenses = [];
+import { getVisibleExpenses } from './selectors/visibleExpenses.js';
 
-const defaultFilters = {
-    text: 'default',
-    sortBy: 'amount/date',
-    startDate: undefined,
-    endDate: undefined
-}
+import { configureStore } from './store/configureStore.js';
 
-// Expenses Action Generators
-
-const AddExpense = ( {description = 'default text', note = 'default note', price = 123 } = {}) => ({
-    type: 'ADD_EXPENSE',
-    expense: {
-        id: uuid(),
-        description,
-        note,
-        price,
-        createdAt: 112233
-    }
-})
-
-const RemoveExpense = (id) => ({
-    type: 'REMOVE_EXPENSE',
-    id: id
-})
-
-const EditExpense = (id, updated_data ) => ({
-    type: 'EDIT_EXPENSE',
-    id,
-    updated_data
-})
-
-const expensesReducer = (state = defaultExpenses, action) => {
-    switch(action.type) {
-        case 'ADD_EXPENSE':
-            return [...state, action.expense ]
-        case 'REMOVE_EXPENSE':
-            return state.filter( ({ id }) => id !== action.id );
-        case 'EDIT_EXPENSE':
-            return state.map((expense) => {
-                if(expense.id == action.id) {
-                    return { ...expense, ...action.updated_data }
-                } else {
-                    return expense;
-                }
-            });
-        default: 
-            return state;
-    }
-}
-
-
-const changeTextFilter = (search_data) => ({
-    type: 'TEXT_FILTER',
-    search_data
-})
-
-const sortByDate = () => ({
-    type: 'SORT_BY_DATE',
-    sortBy: 'date'
-})
-
-const sortByAmount = () => ({
-    type: 'SORT_BY_AMOUNT',
-    sortBy: 'amount'
-})
-
-const setStartDate = (startDate) => ({
-    type: 'SET_START_DATE',
-    startDate
-})
-
-const setEndDate = (endDate) => ({
-    type: 'SET_END_DATE',
-    endDate
-})
-
-
-
-const filtersReducer = ( state = defaultFilters, action ) => {
-    switch(action.type) {
-        case 'TEXT_FILTER':
-            return { ...state, ...action.search_data }
-        case 'SORT_BY_DATE':
-            return { ...state, sortBy: action.sortBy}
-        case 'SORT_BY_AMOUNT':
-            return { ...state, sortBy: action.sortBy }
-        case 'SET_START_DATE':
-            return { ...state, startDate: action.startDate }
-        case 'SET_END_DATE':
-            return { ...state, endDate: action.endDate }
-        default:
-            return state;
-    }
-}
-
-
-const store = createStore(combineReducers({expenses: expensesReducer, filters: filtersReducer }));
-
-
-const getVisibleExpenses = (expenses, filters) => {
-    return expenses.filter((expense) => {
-        const startDateMatch = expense.createdAt >= filters.startDate;
-        const textMatch = expense.description.includes(filters.text);
-
-        return startDateMatch && textMatch
-    })
-}
+const store = configureStore();
 
 
 store.subscribe(() => {
@@ -136,5 +30,5 @@ store.subscribe(() => {
 // store.dispatch(setStartDate(1234));
 // // store.dispatch(setEndDate(4321));
 
-store.dispatch(AddExpense());
-store.dispatch(setStartDate(123))
+// store.dispatch(AddExpense());
+// store.dispatch(setStartDate(123))
